Rename chartdata folder
