+++ src/index.js
@@ -1,2 +1,1 @@
-export {default as ChartData} from './records/ChartData';
-export {default as ChartDataSeries} from './records/ChartDataSeries';
+export {default as ChartData} from './chartdata/ChartData';
